+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add `joint_state_desired` publisher to `franka_gazebo`
 * No further ROS Kinetic support, since [End-of-Life was in April 2021](http://wiki.ros.org/Distributions)
 * Make position + orientation targets threadsafe in cartesian example controller
+* Add singularity warning to `franka_gazebo` if Jacobian becomes singular
 
 ## 0.8.1 - 2021-09-08